// src/app/components/excent-text/excent-text.types.ts
export type ExcentTextVariant =
  | 'display'
  | 'title'
  | 'subtitle'
  | 'body'
  | 'footer-text'
  | 'kicker'

export type ExcentTextAs =
  | 'h1' 
  | 'h2' 
  | 'h3' 
  | 'h4' 
  | 'h5' 
  | 'h6'
  | 'p' 
  | 'span' 
  | 'div' 
  | 'small'

export type ExcentTextColor =
  | 'primary'
  | 'secondary'
  | 'muted'
  | 'inherit'